Consolidate chat dashboard components into shared hook
Extracts all duplicated logic (GET_CHAT_BY_ID query, CREATE_MESSAGE
mutation, SUBSCRIBE_TO_CHAT subscription, message dedup+sort, isSending
state, sendMessage) into useChatDashboard(senderType). Both dashboard
components are now ~35 lines each and differ only in their context hook,
senderType, and which field they use for the other participant's name.
Also adds themed loading spinner and error Alert to ChatDashboardUser to
match the improvements already made to the Member side.
